stat: split ProcessStat by adding VideoStat files
